test: add e2e tests for `watchFiles` option
